requirements.txt
google-genai
python-dotenv
PySide6
